LayoutTests/http/tests/security/xssAuditor/xss-filter-bypass-long-string-expected.txt

CONSOLE MESSAGE: line 79: Refused to execute a JavaScript script. Source code of script found within request.
